requirements.txt
streamlit
pillow
torch
transformers
huggingface_hub
ollama
